inferTypeVariables: handle Object inferFrom
This situation comes up often when calling something along the lines of:

  ops.op("name").inType(Double.class).apply()

which creates an OpRef looking for a

  Function<Double, Object>

Suppose the Op we wanted to match was a

  Function<I, Iterable<O>>

  (Disregard how bad of a Function typing this is)

MatchingUtils.inferTypeVariables() did not put the O in its
typeMappings, which throws IAEs later on. This is specifically because
Object IS NOT a superType of Iterable (and therefore I do not think this
problem can occur with any type other than Object??).

This problem is solved by adding Anys whenever an Object is found.