Test OpBuilder specifically without input
This test should be differentiated from the no output tests within
OpBuilderTest in that tests within this class are more complex. The
no-output tests within OpBuilderTest merely ensure that a) the methods
continue to exist and b) that basic functionality is available. This
class tests more complicated functionality contingent on no output
(which is currently limited to ensuring ParameterizedType outputs do not
 cause error when no output is specified)